dialogue/views.py
import django_filters
from rest_framework import viewsets, filters

from .models import Question, Answer, Keyword
from .serializer import QuestionSerializer, AnswerSerializer, KeywordSerializer


class QuestionViewSet(viewsets.ModelViewSet):
    queryset = Question.objects.all()
    serializer_class = QuestionSerializer


class AnswerViewSet(viewsets.ModelViewSet):
    queryset = Answer.objects.all()
    serializer_class = AnswerSerializer


class KeywordViewSet(viewsets.ModelViewSet):
    queryset = Keyword.objects.all()
    serializer_class = KeywordSerializer
