requirements.txt
Django~=2.1
PyMySQL
djangorestframework
django-filter
django-mysql
